+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# next
+# 1.3.2
 - Improvements in extensions organization.
 - Update Flame to `1.0.0-releasecandidate.13` version.
 - Replace `HasGameRef` for the own `BonfireHasGameRef`.